Update translate capturing algorithm
